Refactor post handling and UI components for improved functionality and clarity
