+++ app/Http/Controllers/Private/PostController.php
@@ -10,7 +10,6 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Post\CreatePostRequest;
 use App\Http\Requests\Post\UpdatePostRequest;
 use App\Http\Resources\PostResource;
-use App\Http\Resources\PostIndexResource;
 use App\Models\Post;
 use Inertia\Inertia;
 
@@ -36,7 +35,7 @@ class PostController extends Controller
             ->orderBy('created_at','desc');
 
         if ($user->hasPermission('post.list')) {
-            return PostIndexResource::collection($query->paginate(10));
+            return PostResource::collection($query->paginate(10));
         }
 
         if ($user->hasPermission('post.list.own')) {
@@ -46,7 +45,7 @@ class PostController extends Controller
             });
         }
 
-        return PostIndexResource::collection($query->paginate(10));
+        return PostResource::collection($query->paginate(10));
     }
 
     public function showPost(Post $post)
