+++ resources/js/ui/widgets/private/grid/PostGrid.svelte
@@ -8,6 +8,8 @@
 
     $: ({ posts } = $page.props);
 
+    $:console.log(posts)
+
     let can = {
         update: hasPermission("post.update"),
         deactivate: hasPermission("post.deactivate"),
